clinic_manager/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.1
colorama==0.4.6
cryptography==42.0.5
DateTime==5.5
defusedxml==0.8.0rc2
Django==5.0.4
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.3.7
MarkupSafe==2.1.5
momo==0.1.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.2.0
sh==2.0.6
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
zope.interface==6.2
